Revert "implement alfred.run()"
This reverts commit 0beb9216fce9ef78eb71e7c7ff049cae35823561.
