+++ src/alfred/helpmates/localserver.py
@@ -1,6 +1,5 @@
 from builtins import map
 from builtins import object
-from builtins import callable as builtins_callable
 from flask import Flask, send_file, redirect, url_for, abort, request, make_response, session
 import re
 from uuid import uuid4
@@ -17,11 +16,6 @@ class C(object):
     experiment = None
 
 
-class Generator(object):
-    def generate_experiment(self):
-        pass
-
-
 script = C()
 script.experiment = None
 script.generator = None
@@ -32,12 +26,7 @@ def set_experiment(exp):
 
 
 def set_generator(generator):
-    if builtins_callable(generator):
-        print(generator)
-        script.generator = Generator()
-        script.generator.generate_experiment = generator.__get__(script.generator, Generator)
-    elif callable(generator.generate_experiment):
-        script.generator = generator
+    script.generator = generator
 
 
 @app.route('/start', methods=['GET', 'POST'])
