+++ src/script.py
@@ -10,7 +10,7 @@ from alfred.section import SegmentedSection
 # from alfred.layout import *
 # from alfred.helpmates import *
 
-from alfred import Experiment, run
+from alfred import Experiment
 
 #################################################
 # - Section 2: Global variables and functions - #
@@ -25,7 +25,7 @@ from alfred import Experiment, run
 ########################################
 
 
-def generate_experiment(self):
+def generate_experiment():
     exp = Experiment()
 
     # --- Page 1 --- #
@@ -39,6 +39,3 @@ def generate_experiment(self):
     exp.page_controller.append(main)
 
     return exp
-
-
-run(generate_experiment)
